Small updates to remove warnings, mainly making casts explicit

+++ Computer Graphics From Scratch/Dielectric.cpp
@@ -30,5 +30,5 @@ float Dielectric::Reflectance(float cosine, float refIdx)
 	// Schlick's approximation for reflectance
 	float r0 = (1 - refIdx) / (1 + refIdx);
 	r0 *= r0;
-	return r0 = (1 - r0) * pow((1 - cosine), 5);
+	return r0 = (1 - r0) * powf((1 - cosine), 5);
 }
